Chore: Update image paths in admin stylesheets
WATER-2140

Uses gulp-replace to change any image paths from /public/images to
/public/admin/images.

This seems to be possible via a SASS global variable but this application
has no custom CSS or an entry SCSS file so opted to replace the text
instead.

Because there is not any SASS the gulp-sass and gulp-sourcemaps deps
could be removed as well.